chore: fix build minor refactor
