+++ src/util/constants.ts
@@ -116,3 +116,16 @@ export const compareLandmarksChanged: (prevLandmarks: any[], newLandmarks: any[]
 
 export const CYCLE_PANEL_MESSAGE = `[spacebar] to cycle through views`
 export const BS_ENDPOINT = new InjectionToken<string>('BS_ENDPOINT')
+
+export const UNSUPPORTED_PREVIEW = [{
+  text: 'Preview of Colin 27 and JuBrain Cytoarchitectonic',
+  previewSrc: './res/image/1.png',
+}, {
+  text: 'Preview of Big Brain 2015 Release',
+  previewSrc: './res/image/2.png',
+}, {
+  text: 'Preview of Waxholm Rat V2.0',
+  previewSrc: './res/image/3.png',
+}]
+
+export const UNSUPPORTED_INTERVAL = 7000
